Encapsulate MetricsExportPanel
