+++ public/components/metrics/top_menu/top_menu.tsx
@@ -5,13 +5,11 @@
 
 import {
   EuiButton,
-  EuiButtonEmpty,
   EuiComboBoxOptionOption,
   EuiFieldText,
   EuiFlexGroup,
   EuiFlexItem,
   EuiPopover,
-  EuiPopoverFooter,
   EuiSelect,
   EuiSpacer,
   EuiSuperDatePicker,
@@ -27,10 +25,6 @@ import { resolutionOptions } from '../../../../common/constants/metrics';
 import { MetricType } from '../../../../common/types/metrics';
 import { uiSettingsService } from '../../../../common/utils';
 import SavedObjects from '../../../services/saved_objects/event_analytics/saved_objects';
-import { getSavedObjectsClient } from '../../../services/saved_objects/saved_object_client/client_factory';
-import { OSDSavedVisualizationClient } from '../../../services/saved_objects/saved_object_client/osd_saved_objects/saved_visualization';
-import { addMultipleVizToPanels, uuidRx } from '../../custom_panels/redux/panel_slice';
-import { sortMetricLayout, updateMetricsWithSelections } from '../helpers/utils';
 import {
   allAvailableMetricsSelector,
   metricsLayoutSelector,
@@ -87,15 +81,9 @@ export const TopMenu = ({
   const dispatch = useDispatch();
   const allAvailableMetrics = useSelector(allAvailableMetricsSelector);
   const handleAddMetric = (metric: any) => dispatch(selectMetric(metric));
-  const metricsLayout = useSelector(metricsLayoutSelector);
-  const sortedMetricsLayout = sortMetricLayout([...metricsLayout]);
 
-  const [visualizationsMetaData, setVisualizationsMetaData] = useState<any>([]);
   const [originalPanelVisualizations, setOriginalPanelVisualizations] = useState<MetricType[]>([]);
   const [isSavePanelOpen, setIsSavePanelOpen] = useState(false);
-  const [selectedPanelOptions, setSelectedPanelOptions] = useState<
-    Array<EuiComboBoxOptionOption<unknown>> | undefined
-  >([]);
 
   // toggle between panel edit mode
   const editPanel = (editType: string) => {
@@ -168,67 +156,6 @@ export const TopMenu = ({
     </EuiButton>
   );
 
-  const handleSavingObjects = async () => {
-    let savedMetricIds = [];
-
-    try {
-      savedMetricIds = await Promise.all(
-        sortedMetricsLayout.map(async (metricLayout, index) => {
-          const updatedMetric = updateMetricsWithSelections(
-            visualizationsMetaData[index],
-            startTime,
-            endTime,
-            spanValue + resolutionValue
-          );
-
-          if (metricLayout.metricType === 'prometheusMetric') {
-            return OSDSavedVisualizationClient.getInstance().create(updatedMetric);
-          } else {
-            return getSavedObjectsClient({
-              objectId: metricLayout.id,
-              objectType: 'savedVisualization',
-            }).update({
-              ...updatedMetric,
-              objectId: metricLayout.id,
-            });
-          }
-        })
-      );
-    } catch (e) {
-      const message = 'Issue in saving metrics';
-      console.error(message, e);
-      setToast(message, 'danger');
-      return;
-    }
-
-    setToast('Saved metrics successfully!');
-
-    if (selectedPanelOptions.length > 0) {
-      try {
-        const allMetricIds = savedMetricIds.map((metric) => metric.objectId);
-        const soPanels = selectedPanelOptions.filter((panel) => uuidRx.test(panel.panel.id));
-        const opsPanels = selectedPanelOptions.filter((panel) => !uuidRx.test(panel.panel.id));
-
-        dispatch(addMultipleVizToPanels(soPanels, allMetricIds));
-        const savedMetricsInOpsPanels = await Promise.all(
-          opsPanels.map((panel) => {
-            return http.post(`${CUSTOM_PANELS_API_PREFIX}/visualizations/multiple`, {
-              body: JSON.stringify({
-                panelId: panel.panel.id,
-                savedVisualizationIds: allMetricIds,
-              }),
-            });
-          })
-        );
-      } catch (e) {
-        const message = 'Issue in saving metrics to panels';
-        console.error(message, e);
-        setToast('Issue in saving metrics', 'danger');
-      }
-      setToast('Saved metrics to Dashboards successfully!');
-    }
-  };
-
   return (
     <>
       <EuiFlexGroup gutterSize="s">
@@ -277,38 +204,12 @@ export const TopMenu = ({
             closePopover={() => setIsSavePanelOpen(false)}
           >
             <MetricsExportPanel
-              http={http}
-              visualizationsMetaData={visualizationsMetaData}
-              setVisualizationsMetaData={setVisualizationsMetaData}
-              sortedMetricsLayout={sortedMetricsLayout}
-              selectedPanelOptions={selectedPanelOptions}
-              setSelectedPanelOptions={setSelectedPanelOptions}
+              setIsSavePanelOpen={setIsSavePanelOpen}
+              startTime={startTime}
+              endTime={endTime}
+              resolutionValue={resolutionValue}
+              spanValue={spanValue}
             />
-            <EuiPopoverFooter>
-              <EuiFlexGroup justifyContent="flexEnd">
-                <EuiFlexItem grow={false}>
-                  <EuiButtonEmpty
-                    size="s"
-                    onClick={() => setIsSavePanelOpen(false)}
-                    data-test-subj="metrics__SaveCancel"
-                  >
-                    Cancel
-                  </EuiButtonEmpty>
-                </EuiFlexItem>
-                <EuiFlexItem grow={false}>
-                  <EuiButton
-                    size="s"
-                    fill
-                    onClick={() => {
-                      handleSavingObjects().then(() => setIsSavePanelOpen(false));
-                    }}
-                    data-test-subj="metrics__SaveConfirm"
-                  >
-                    Save
-                  </EuiButton>
-                </EuiFlexItem>
-              </EuiFlexGroup>
-            </EuiPopoverFooter>
           </EuiPopover>
         </EuiFlexItem>
       </EuiFlexGroup>
